Update types to include tag and hash code
It's reasonable to rely on an external module for bignums, because it's
reasonable to implement bignums with BigInt on the browser or an
external wasm module otherwise.  However hashq can't nicely be
implemented in a non-browser context, so we might as well make it our
own.  The nice thing about this is that we get a mutability bit for
vectors and such-like; WebAssembly's notion of mutability is useful but
not exactly the same as what we need for Scheme.