Enhance compiler to make a <wasm> module
This compiler enhancement makes it so that the compiler pulls in the
different parts of the run-time as needed.  For example we should only
have a symbol table if the module uses symbols.  We get to the point of
dumping the wasm but we can't assembly it yet because the resolver
doesn't yet know about reftypes.