Fix vars defined by joins
In a funny interaction between WebAssembly's type system, the Beyond
Relooper algorithm, and Guile's CPS soup, we were producing invalid
wasm.  The thing is that in Wasm you can have non-nullable locals but
they have to be defined before first use, otherwise they don't have a
default value.  We were respecting that condition but not in such a way
that it was obvious to WebAssembly's deliberately limited flow analysis.
For the time being, we now explicitly initialize such join vars to a
default value at the head of the function.  This takes space but no time
at higher compilation levels, as this is strictly a dead store.