Better diagnostics on not-yet-implemented stdlib, primcall lowering
Now when the stdlib dies because something isn't implemented, it will
print a reason first.  That will let you know what to implement.

Similarly it would seem that with new Guile, perhaps due to the `box'
primcall change recently, we start to see a new set of primcalls.
Update our lower-primcalls pass to catch all of these.  Fix scm->u64,
which was formerly accepting negative numbers.